# Create a JMenuItem object to represent menu items in the JMenu API