--- a/libraries/cms/menu/site.php
+++ b/libraries/cms/menu/site.php
@@ -91,7 +91,7 @@ class JMenuSite extends JMenu
 					// Set the query
 					$db->setQuery($query);
 
-					return $db->loadObjectList('id');
+					return $db->loadObjectList('id', 'JMenuItem');
 				},
 				array(),
 				md5(get_class($this)),