Pull Request for Issue #13054

### Summary of Changes

Menu items as loaded by `JMenuSite` (as this is the only `JMenu` class actually loading anything) will now be loaded into this new `JMenuItem` object instead of a plain PHP object.  The object's properties are based on the query used by `JMenuSite` instead of a direct mapping to the `#__menu` database table as this is actually what is used in runtime.

All of the class properties are public (emulating the existing behavior) except for the parameters.  This is a protected field and through a magic getter and setter still allow read/write access to this field.  This is required to allow the menu item parameters to be "lazy loaded" into a Registry instance only when the parameters are actually needed.  Thus, the auto-conversion of all parameters in `JMenu::load()` is removed.

The magic getter and setter, and extending from `JObject`, are deprecated and should be removed with 4.0.  These measures are in place now to retain B/C with the existing plain PHP objects that are currently used by the API.  The parameters field has a dedicated getter and setter method which should be used instead and we should discourage adding additional undocumented properties to this object going forward.

### Testing Instructions

When the site's menu tree is loaded into `JMenu`, the data should continue to be loaded correctly but instead of being mapped to a `stdClass` object, each item in the `$_items` array should now be a `JMenuItem` object.  The `params` field should stay in the form of a JSON string until the menu item's parameters are explicitly read, at which time it will be converted to a `Registry` object.

### Documentation Changes Required

N/A