@ssnobben 2016-12-03
Thanks Michael!. 

A side note to this large menu performance and so I hijack this for an idea about modules too. 

As now when you have many site languages and for example you have one and several modules that render info data from a component you have a site design with same module in the same positions with the same content data but with different language heading/text/title/description before/after etc then you have to set up new modules for each of these languages that are exactly the same? 

Let me know if I have missed something that you can do different to that problem instead of creating a copy of every module of your page, write a new module title to every languages you have on your Joomla site. 

So what it means is that if you have 7 languages you have to create 7 different modules just for changing the title to map to each language instead of just have one module that can do exactly the same but with different language title text.

What I would like to have is o n e module that is in the same position with the same data input but where you can have language settings of Titles for a ”master module” like: 

Title (fr-FR): French title of module
Title (es-ES): Spanish title of module
Title (it-IT): Italian title of module
Etc 

Wouldn’t that be better and maybe also for performance bcs the module can be cached but the extra title text can be dynamically updated for each language? 

Advanced module mgr doesn’t have this feature. Maybe some modules can have this option(s) for being a master module with these settings?