@ggppdk 2016-12-03
(Database is on SSD, and duplicating menu items via PHP code to go up to 6000 menu items)

**load()   (on a JMenuSite object)**
-- Before patch: 2.710 seconds, With 600 parameters for 6000 menu items
-- Before patch: 0.498 seconds, With <100 parameters for 6000 menu items
-- After patch: 0.045 seconds
**Speedup 10x - 60x**  

----

Also since languageFilter system plugin **onAfterInitialise** does:
$router = $this->app->getRouter();
- which eventually loads all menu items (and their parameters)

**onAfterInitialise times**
Before patch: 2,742 seconds, with 600 parameters per item for 6000 items
Before patch: 0.625 seconds, with<100 parameters per item for 6000 items
After patch: 0.181 seconds

----

Please note that if you display all 6000 menu items in large menu in your page 
or if you display a page with all menus
- then all parameters will be forced to be loaded, and you will get no benefits