@ggppdk 2016-12-03
I have tested this item :white_check_mark: successfully on 20c83ce459f0340f37d54dc800e6b40e92ae9e7f<br /><br />Tested using PHP 7.0.13<hr /><sub>This comment was created with the <a href="https://github.com/joomla/jissues">J!Tracker Application</a> at <a href="https://issues.joomla.org/tracker/joomla-cms/13073">issues.joomla.org/tracker/joomla-cms/13073</a>.</sub>