@alikon 2016-12-06
I have tested this item :white_check_mark: successfully on 20c83ce459f0340f37d54dc800e6b40e92ae9e7f<br /><br />
p.s.
@ggppdk can you share your script  for create tons of menu items?<hr /><sub>This comment was created with the <a href="https://github.com/joomla/jissues">J!Tracker Application</a> at <a href="https://issues.joomla.org/tracker/joomla-cms/13073">issues.joomla.org/tracker/joomla-cms/13073</a>.</sub>