@jeckodevelopment 2016-12-06
RTC<hr /><sub>This comment was created with the <a href="https://github.com/joomla/jissues">J!Tracker Application</a> at <a href="https://issues.joomla.org/tracker/joomla-cms/13073">issues.joomla.org/tracker/joomla-cms/13073</a>.</sub>