@ggppdk 2016-12-06
This change should benefit all AJAX requests too,
since any getMenu() called as a result of some code inside onAfterInitialize event
will no longer cause a slow down on website with many menu items